+++ packages/article-byline/__tests__/android/article-byline-with-links-style.android.test.js
@@ -1,4 +1,0 @@
-import { ArticleBylineWithLinks } from "../../src/article-byline";
-import shared from "../shared-with-styles.native";
-
-shared(ArticleBylineWithLinks);