+++ packages/article-comments/__tests__/android/shared-with-style.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-with-style";
-
-shared();