+++ packages/article-list/__tests__/ios/article-list-with-styles.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-with-styles.native";
-
-shared();