+++ packages/article-main-comment/__tests__/android/article-internal-components.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-internal-components.native";
-
-shared();