+++ packages/article-skeleton/__tests__/android/body-utils.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../body-utils.native";
-
-shared();