+++ packages/pagination/__tests__/ios/pagination-wrapper.ios.test.js
@@ -1,4 +1,0 @@
-import shared from "../shared.native";
-import withPageState from "../../src/pagination-wrapper";
-
-shared(withPageState);