+++ packages/related-articles/__tests__/ios/std-0articles.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-std-0articles.native";
-
-shared();