+++ packages/slice-layout/__tests__/ios/clac.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../clac.native";
-
-shared();