+++ packages/video-label/__tests__/ios/video-label-with-style.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../video-label-with-style.native";
-
-shared();