+++ packages/video/__tests__/ios/video.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../video-shared.native";
-
-shared();