+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.37](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.5.36...@times-components/article-byline@3.5.37) (2021-08-09)
+
+**Note:** Version bump only for package @times-components/article-byline
+
+
+
+
+
 ## [3.5.36](https://github.com/newsuk/times-components/compare/@times-components/article-byline@3.5.35...@times-components/article-byline@3.5.36) (2021-07-05)
 
 **Note:** Version bump only for package @times-components/article-byline