+++ packages/tealium/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.3](https://github.com/newsuk/times-components/compare/@times-components/tealium@1.14.2...@times-components/tealium@1.14.3) (2021-08-09)
+
+**Note:** Version bump only for package @times-components/tealium
+
+
+
+
+
 ## [1.14.2](https://github.com/newsuk/times-components/compare/@times-components/tealium@1.14.1...@times-components/tealium@1.14.2) (2021-05-27)
 
 **Note:** Version bump only for package @times-components/tealium